# Conflicts:
#	gamepadROSPUB.py

+++ gamepadROSPUB.py
@@ -23,8 +23,8 @@ def callback(data):
     pub.publish(twist)
     # print(str(Twist))
     # FORWARD/BACK =  1/-1.
-    if Float32(twist.linear.x) < (-0.01):
-        if Float32(twist.linear.x) < (-0.99):
+    if Float32(twist.linear.x) < Float32(-0.01):
+        if Float32(twist.linear.x) < Float32(-0.99):
             hold1 = ('Backwards   ' + str(Float32(twist.linear.x)) + '           ')
         hold1 = ('Backwards   ' + str(Float32(twist.linear.x)))
     elif twist.linear.x > (0):
@@ -34,8 +34,8 @@ def callback(data):
     else:
         hold1 = ('Stop        ' + str(Float32(twist.linear.x)) + '           ')
     # LEFT/RIGHT = 1/-1
-    if Float32(twist.angular.z) < (-0.01):
-        if Float32(twist.angular.z) < (-0.99):
+    if Float32(twist.angular.z) < Float32(-0.01):
+        if Float32(twist.angular.z) < Float32(-0.99):
             hold2 = ('Right       ' + str(Float32(twist.angular.z)) + '           ')
         hold2 = ('Right       ' + str(Float32(twist.angular.z)))
     elif twist.angular.z > (0):
